pyproject.toml
[project]
name = "llm-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nazer, Mohamed Ilyan", email = "mohamed.ilyan+Boskalis@boskalis.com" }
]
requires-python = ">=3.12"
dependencies = [
    "gensim>=4.3.2",
    "gradio>=5.45.0",
    "ipywidgets>=8.1.7",
    "jupyterlab>=4.4.7",
    "langchain[docarray]>=0.3.27",
    "langchain-chroma>=0.2.6",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.33",
    "matplotlib>=3.10.6",
    "numpy>=1.26.4",
    "openai>=1.107.1",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.13.1",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.1",
    "datasets>=4.0.0",
    "sentencepiece>=0.2.1",
    "google-generativeai>=0.8.5",
    "anthropic>=0.67.0",
    "unstructured>=0.18.14",
    "chromadb>=1.0.21",
    "plotly>=6.3.0",
    "jupyter-dash>=0.4.2",
    "beautifulsoup4>=4.13.5",
    "pydub>=0.25.1",
    "modal>=1.1.4",
    "ollama>=0.5.3",
    "accelerate>=1.10.1",
    "bitsandbytes>=0.47.0",
    "psutil>=7.0.0",
    "setuptools>=80.9.0",
    "speedtest-cli>=2.1.3",
    "sentence-transformers>=5.1.0",
    "feedparser>=6.0.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/llm_playground"]

